Remove date function duplication
The issue here is less with the code duplication itself and more with
the duplicative tests required; a problem which will only get worse as
we add more date-related functions.
